// Angular,TypeScript(FrontEnd)/ovoWHM_web/src/app/modules/Masters/flock/flock.type.ts
export interface flockdetails{
     Id:number,
     Name:string,
     Area:number,
     Location:string,
     CreatedDate: Date;
     CreatedByUserId: number;
     UpdatedDate: Date;
     UpdatedByUserId: number;
}
export interface flockitems{
    Id:number,
    Name:string,
    Area:number,
    Location:string,
    CreatedDate: Date;
    CreatedByUserId: number;
    UpdatedDate: Date;
    UpdatedByUserId: number;
}
  